[move] 기존 custom 어노테이션 패키지 이동 (#277)
